@@ -1,18 +1,29 @@
 <?php
 
+namespace App\Controller;
+
+
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-class StudentController extends AbstractController{
 
-/**
- * @Route("/index",name = "index")
- */
-#[Route('/index', name: 'index')]
-  public function index()
+
+
+
+
+class StudentController extends AbstractController
+{
+    #[Route('/student', name: 'app_student')]
+    public function index()
  {
-    return new Response('Bonjour mes étudiants') ;
- }
 
+    return new Response("bonjour mes etudiants");
+
+ }
 }
+
 ?>
+
+
+
